requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
filelock==3.20.1
fsspec==2025.12.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.3.5
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
scikit-learn==1.8.0
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.9.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
websocket-client==1.9.0
